FreeBSD: `clippy --fix`

+++ src/columns/start_time.rs
@@ -96,7 +96,7 @@ impl Column for StartTime {
 impl Column for StartTime {
     fn add(&mut self, proc: &ProcessInfo) {
         let start_time = Local
-            .timestamp_opt(proc.curr_proc.ki_start.tv_sec as i64, 0)
+            .timestamp_opt(proc.curr_proc.ki_start.tv_sec, 0)
             .unwrap();
         let raw_content = start_time;
         let fmt_content = format_start_time(raw_content);
